BTC testnet/mainnet test pass (#18)
