+++ src/crypto/bip44.ts
@@ -1,16 +1,18 @@
-const bip44 = {
-  createDerivePath(network) {
-    /*
-    In fact, not every testnet of coins has an index of 1
-    Therefore, specify the testnet coin index in the settings
-    */
+type Path = string
 
-    const { coinIndex } = network.settings.bip44
-    const addressIndex = 0
+const createDerivePath = ({ coinIndex, addressIndex }: {
+  coinIndex: number,
+  addressIndex: number
+}): Path => {
+  /*
+  In fact, not every testnet of coins has an index of 1
+  Therefore, specify the testnet coin index in the settings
+  */
 
-    const path = `m/44'/${coinIndex}'/0'/0/${addressIndex}`
-    return path
-  }
+  const path = `m/44'/${coinIndex}'/0'/0/${addressIndex}`
+  return path
 }
 
-export default bip44
+export default {
+  createDerivePath
+}
